+++ examples/shadow/main.rs
@@ -1,6 +1,7 @@
 extern crate gaclen;
 
 mod shaders;
+mod geometry;
 
 use gaclen::graphics;
 
@@ -24,7 +25,7 @@ fn main() {
 	let mut events_loop = EventsLoop::new();
 	let window = std::sync::Arc::new(
 		WindowBuilder::new()
-			.with_title("Quad example")
+			.with_title("Shadowing example")
 			.with_dimensions((1280, 720).into())
 			.with_min_dimensions((1280, 720).into())
 			.build(&events_loop).unwrap()
@@ -34,22 +35,15 @@ fn main() {
 	let mut device = graphics::device::Device::new(&context, window.clone(), graphics::device::PresentMode::Immediate).unwrap();
 	println!("Initialized device: {:?}", device);
 
-	let pass = {
-		let vs = shaders::vertex::Shader::load(device.logical_device()).unwrap();
-		let fs = shaders::fragment::Shader::load(device.logical_device()).unwrap();
+	let albedo_pass = {
+		let vs = shaders::albedo::vertex::Shader::load(device.logical_device()).unwrap();
+		let fs = shaders::albedo::fragment::Shader::load(device.logical_device()).unwrap();
 
 		graphics::pass::AlbedoPass::new::<_, _, Vertex>(&device, vs.main_entry_point(), (), fs.main_entry_point(), ()).unwrap()
 	};
 
-	let triangle_buffer = device.create_buffer([
-		Vertex { position: [-0.5, 0.5, 0.0 ], color: [ 0.25, 0.75, 0.25, 1.0 ] },
-		Vertex { position: [ 0.5,-0.5, 0.0 ], color: [ 0.75, 0.25, 0.25, 1.0 ] },
-		Vertex { position: [ 0.5, 0.5, 0.0 ], color: [ 0.75, 0.75, 0.25, 0.0 ] },
-
-		Vertex { position: [-0.5,-0.5, 0.0 ], color: [ 0.0, 0.0, 0.0, 1.0 ] },
-		Vertex { position: [ 0.5,-0.5, 0.0 ], color: [ 1.0, 0.0, 0.0, 1.0 ] },
-		Vertex { position: [-0.5, 0.5, 0.0 ], color: [ 0.0, 1.0, 0.0, 1.0 ] },
-	].iter().cloned()).unwrap();
+	let quad = geometry::generate_quad(&device);
+	let cube = geometry::generate_quad(&device);
 
 	let mut recreate_swapchain = false;
 
@@ -69,11 +63,11 @@ fn main() {
 		}
 
 		let clear_color = [0.0, 0.0, 0.0, 1.0];
-		let push_constants = push_constants_from_time(start_time.elapsed().as_secs_f32(), window.get_inner_size().unwrap().into());
 
 		let after_frame = device.begin_frame().unwrap()
-			.begin_pass(&pass, vec![clear_color.into()])
-				.draw(vec![triangle_buffer.clone()], push_constants)
+			.begin_pass(&albedo_pass, vec![clear_color.into()])
+				.draw(vec![quad.clone()], ())
+				.draw(vec![cube.clone()], ())
 				.finish_pass()
 			.finish_frame();
 		
@@ -100,24 +94,4 @@ fn main() {
 	let fps: f64 = frame_count as f64 / run_duration;
 
 	println!("Produced {} frames over {:.2} seconds ({:.2} avg fps)", frame_count, run_duration, fps);
-}
-
-fn push_constants_from_time(time: f32, window_resolution: (u32, u32)) -> shaders::vertex::ty::PushConstantData {
-	let time = time / 5.0;
-
-	let x = time.cos();
-	let y = time.sin();
-
-	let view = cgmath::Matrix4::look_at(
-		cgmath::Point3 { x, y, z: 1.0 },
-		cgmath::Point3 { x: 0.0, y: 0.0, z: 0.0 },
-		cgmath::Vector3 { x: 0.0, y: 0.0, z: 1.0 });
-	
-	let aspect = window_resolution.0 as f32 / window_resolution.1 as f32;
-	
-	let proj: cgmath::Matrix4<f32> = cgmath::PerspectiveFov { fovy: cgmath::Deg(60.0).into(), aspect, near: std::f32::EPSILON, far: 4.0 }.into();
-
-	let mvp = proj * view;
-
-	shaders::vertex::ty::PushConstantData { MVP: mvp.into() }
 }