+++ src/lib.rs
@@ -13,8 +13,6 @@
 
 /// used for hardware acceleration.
 pub use vulkano;
-/// used for compiling shaders.
-pub use vulkano_shaders;
 /// used for OS integration.
 pub use winit;